+++ src/components/transactions-table/components/transaction-age/transaction-age.tsx
@@ -8,15 +8,18 @@ import { DataTestIdsEnum } from 'constants/dataTestIds.enum';
 })
 export class TransactionAge {
   @Prop() age: string;
+  @Prop() class?: string = 'transaction-age';
   @Prop() tooltip?: string;
 
   render() {
     const component = this.tooltip ? (
-      <span title={this.tooltip} data-testid={DataTestIdsEnum.transactionAge}>
+      <span title={this.tooltip} data-testid={DataTestIdsEnum.transactionAge} class={this.class}>
         {this.age}
       </span>
     ) : (
-      <span data-testid={DataTestIdsEnum.transactionAge}>{this.age}</span>
+      <span data-testid={DataTestIdsEnum.transactionAge} class={this.class}>
+        {this.age}
+      </span>
     );
 
     return <span>{component}</span>;